requirements.txt
requests
beautifulsoup4
python-dotenv
line-bot-sdk
schedule
pytz
feedparser
deep-translator
